Created a separate json file for sheet details
